service_python_shared/requirements.txt
opencv-python>=4.11.0.86
aio-pika>=9.0.0
pydantic>=2.0  
orjson>=3.10.18
setuptools<81
python-dotenv>=1.1.0  
loguru>=0.7.3
ruff>=0.12.0
pytest>=8.4.1
pytest-asyncio>=1.0.0
grpcio>=1.73.0
grpcio-tools>=1.73.0
